refactor: Take commented lines of code out
Some script files had commented lines of code which was meat to
testing. This commit take them out.

Also, the .gitconfig core.autocrlf flag was set to false,
as it was causing execution problems with bash don't recognizing
the Windows eol caracter. What this means is that for now on, the
scripts files will be exclusively LF ended.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>