fix: [#1] Fix an export issue with the 'platform-tools' directory
Fix the 'platform-tools' directory export issue.

When the function was trying to export the '~/Android/Sdk/platform-tools'
directory into PATH, if the user hadn't setup Android Studio yet, the function
would fail because this path does not exist until the setup is completed.

This commit addresses this issue with a prompt to the user asking if it wants
to set up Android Studio. Note that we cannot be certain that the user
has, indeed, completed the setup. So one can only assume the best.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>