fix: Fix a variable name causing the git clone command to fail
RCA: When the DOTFILES_PATH variable was being used, its value has to
be accessed by putting a '$' dollar sign in front of it. This was
causing the git clone command to clone the repository into a directory
called DOTFILES_PATH when it should be placing into the variable value.

Solution: Put a '$' sign in fron of the variable to use its value.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>