fix: Fix an inssue preventing the path's creation
RCA: due to how shell scripts works, a missing quote in the if statement
was preventing the correct assertion of the paths's existence and was,
subsequently, falling to creat if it didn't exist.

Sol: Add double quotes to variable FILE_PATH in the if statement. Also
add double quotes to the FILE_PATH value, and use $HOME to exapand the
home directory.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>