fix: Fix creation of authorized_keys file
RCA: The generation of the SSH key requires the creation of an
'authorized_keys' file that contains the user's generate public key. The
steop in which this file was supposed to be created was failing due to a
path error in the code.

SOL: To fix this issue, the path to the public key was modified to be
absolute, instead of relative. This also goes for the authorized_keys
path.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>