fix(#9): Fix error on sourcing scripts
Close #9

RCA: When using associative array in conjunction with an simple array to
mantain order, bash does not recognize the string resulting from the
indexing associative[array[index]].

SOL: Change associative array to simple array. This still maitain the
order of execution when using the -a option and when using custom order
flags.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>